Card refactor and add graphql query for project data
