+++ src/components/projects/Projects.styles.js
@@ -1,15 +1,9 @@
 import styled from "styled-components"
-import { FloatingCard } from "../FloatingCard"
+import { FloatingCard } from "../Card"
 
 export const ProjectsStyled = styled.div`
 display: grid;
 grid-template-columns: 1fr;
 gap: 2rem;
 padding: 3rem;
-
-${FloatingCard} {
-transform: translateY(-50%);
-transition: transform 1000ms ease-in-out;
-
-}
 `
